fix some compatibility issues with 16.04
Google's installer puts a slightly different version of the ops agent
on 16.04 systems and that version is missing some log types and has
different names for a few others.

I anticipated some of this, which is why the config was already a
Jinja2 template rather than a plain file.

So we conditionally tweak the generated config a bit depending on
which Ubuntu release it's getting installed on.

That introduces some potential for breaking the config though, so I
also added a `yamllint` check to the output and some basic tests that
a few of the log types exist/don't exist as appropriate in the
generated config.